Factor out dialog renderer
